rest.control.fabrics.inventory: Refactor routers
# Summary

Refactor  the routers for the following endpoints:

Verb: GET
Path: `/appcenter/cisco/ndfc/api/v1/lan-fabric/rest/control/fabrics/{fabric_name}/inventory/switchesByFabric`

Verb: POST
Path: `/appcenter/cisco/ndfc/api/v1/lan-fabric/rest/control/fabrics/{fabric_name}/inventory/discover`

## Move endpoint handlers

Move handlers to a directory aligned with NDFC endpoint paths

- v1/endpoints/lan_fabric/rest/control/fabrics/inventory/switches_by_fabric_get.py
- v1/endpoints/lan_fabric/rest/control/fabrics/inventory/discover_post.py

## ./app/main.py

- Update imports
- Add routers to app
  - app.include_router(discover_post.router, tags=["Inventory (v1)"])
  - app.include_router(switches_by_fabric_get.router, tags=["Inventory (v1)"])

## Remove unused file(s)

- v1/endpoints/lan_fabric/rest/control/fabrics/inventory.py
