+++ app/v1/endpoints/lan_fabric/rest/control/fabrics/inventory/switches_by_fabric_get.py
@@ -2,13 +2,16 @@
 import copy
 from typing import List
 
-from fastapi import Depends, HTTPException
+from fastapi import APIRouter, Depends, HTTPException
 from sqlmodel import Session, select
 
-from .......app import app
-from .......db import get_session
-from ......models.fabric import Fabric
-from ......models.inventory import SwitchDbModel, SwitchDiscoverBodyModel, SwitchDiscoverItem, SwitchResponseModel
+from ........db import get_session
+from .......models.fabric import Fabric
+from .......models.inventory import SwitchDbModel, SwitchResponseModel
+
+router = APIRouter(
+    prefix="/appcenter/cisco/ndfc/api/v1/lan-fabric/rest/control/fabrics",
+)
 
 
 def build_response_switch(db_switch: SwitchDbModel) -> SwitchResponseModel:
@@ -134,124 +137,6 @@ def build_response_switch(db_switch: SwitchDbModel) -> SwitchResponseModel:
     )
 
 
-def build_db_switch(switch: SwitchDiscoverItem, db_fabric: Fabric) -> SwitchDbModel:
-    """
-    # Summary
-
-    Given a  SwitchDiscoverBodyModel and a Fabric model,
-    return a populated SwitchDbModel object.
-    """
-    return SwitchDbModel(
-        activeSupSlot=1,
-        availPorts=48,
-        ccStatus="",
-        cfsSyslogStatus=0,
-        colDBId=0,
-        connUnitStatus=0,
-        consistencyState=True,
-        contact="",
-        cpuUsage=0,
-        deviceType="Switch_Fabric",
-        displayHdrs="",
-        displayValues="",
-        domain="",
-        domainID=0,
-        elementType="",
-        fabricId=db_fabric.id,
-        fabricName=db_fabric.FABRIC_NAME,
-        fabricTechnology=db_fabric.FF,
-        fcoeEnabled=False,
-        fex=False,
-        fid=0,
-        freezeMode="",
-        health=0,
-        hostName=switch.sysName,
-        index=0,
-        intentedpeerName="",
-        interfaces="",
-        ipAddress=switch.ipaddr,
-        ipDomain="",
-        isEchSupport=False,
-        isLan=False,
-        isNonNexus=False,
-        isPmCollect=False,
-        isSharedBorder=False,
-        isTrapDelayed=False,
-        isVpcConfigured=False,
-        is_smlic_enabled=False,
-        keepAliveState="",
-        lastScanTime=0,
-        licenseDetail="",
-        licenseViolation=False,
-        linkName="",
-        logicalName=switch.sysName,
-        managable=True,
-        mds=False,
-        membership="",
-        mgmtAddress="",
-        memoryUsage=0,
-        mode="Normal",
-        model="",
-        moduleIndexOffset=9999,
-        modelType=0,
-        name=switch.sysName,
-        npvEnabled=False,
-        numberOfPorts=48,
-        operMode=None,
-        operStatus="Minor",
-        peer="",
-        peerlinkState="",
-        peerSerialNumber="",
-        peerSwitchDbId=0,
-        ports=0,
-        present=True,
-        primaryIP="",
-        primarySwitchDbID=0,
-        principal="",
-        protoDiscSettings="",
-        recvIntf="",
-        release=switch.version,
-        role="",
-        sanAnalyticsCapable=False,
-        scope="",
-        secondaryIP="",
-        secondarySwitchDbID=0,
-        sendIntf="",
-        serialNumber=switch.serial_number,
-        sharedBorder=False,
-        sourceInterface="mgmt0",
-        sourceVrf="management",
-        standbySupState=0,
-        status="",
-        switchDbID=None,
-        swType="",
-        swUUID="DCNM-UUID-TEMP",
-        swUUIDId=99999,
-        swWwn="",
-        swWwnName="",
-        sysDescr="",
-        systemMode=None,
-        uid=0,
-        unmanagableCause="",
-        upTime=0,
-        upTimeNumber=0,
-        upTimeStr="",
-        usedPorts=0,
-        username="",
-        vdcId=0,
-        vdcName="",
-        vdcMac="",
-        vendor="cisco",
-        version=switch.version,
-        vpcDomain=0,
-        vrf="management",
-        vsanWwn="",
-        vsanWwnName="",
-        waitForSwitchModeChg=False,
-        wwn="",
-    )
-
-
 def build_success_response():
     """
     # Summary
@@ -265,17 +150,25 @@ def build_success_response():
     return copy.deepcopy(response)
 
 
-@app.get(
-    "/appcenter/cisco/ndfc/api/v1/lan-fabric/rest/control/fabrics/{fabric_name}/inventory/switchesByFabric",
+@router.get(
+    "/{fabric_name}/inventory/switchesByFabric",
     response_model=List[SwitchResponseModel],
 )
-def v1_get_inventory_switches_by_fabric(*, session: Session = Depends(get_session), fabric_name: str):
+def v1_inventory_switches_by_fabric_get(*, session: Session = Depends(get_session), fabric_name: str):
     """
     # Summary
 
-    GET request handler with fabric_name as path parameter.
+    Return a list of switches hosted in fabric_name.
 
-    Return a list of switches for a given fabric_name.
+    ## Endpoint
+
+    ### Verb
+
+    GET
+
+    ### Path
+
+    /appcenter/cisco/ndfc/api/v1/lan-fabric/rest/control/fabrics/{fabric_name}/inventory/switchesByFabric
     """
     db_fabric = session.exec(select(Fabric).where(Fabric.FABRIC_NAME == fabric_name)).first()
     if not db_fabric:
@@ -285,29 +178,4 @@ def v1_get_inventory_switches_by_fabric(*, session: Session = Depends(get_sessio
     if len(db_switches) == 0:
         raise HTTPException(status_code=404, detail=f"No switches found for fabric {fabric_name}")
     response = [build_response_switch(db_switch) for db_switch in db_switches]
-    return response
-
-
-@app.post("/appcenter/cisco/ndfc/api/v1/lan-fabric/rest/control/fabrics/{fabric_name}/inventory/discover")
-def v1_post_inventory_discover(*, session: Session = Depends(get_session), fabric_name: str, switch_discovery_body: SwitchDiscoverBodyModel):
-    """
-    # Summary
-
-    POST request handler with fabric_name as path parameter.
-
-    Discover switches for a given fabric_name.
-    """
-    db_fabric = session.exec(select(Fabric).where(Fabric.FABRIC_NAME == fabric_name)).first()
-    if not db_fabric:
-        raise HTTPException(status_code=404, detail=f"Fabric {fabric_name} not found")
-    fabric_id = db_fabric.id
-    db_switches = session.exec(select(SwitchDbModel).where(SwitchDbModel.fabricId == fabric_id)).all()
-    for db_switch in db_switches:
-        if db_switch.serialNumber in [discovery_body.serial_number for discovery_body in switch_discovery_body.switches]:
-            raise HTTPException(status_code=500, detail=f"Switch {db_switch.serialNumber} already exists in fabric {fabric_name}")
-    for discovery_body in switch_discovery_body.switches:
-        db_switch = build_db_switch(discovery_body, db_fabric)
-        session.add(db_switch)
-    session.commit()
-    response = build_success_response()
     return response
